Pull out encrypted_by_default tests from user_directory tests (#10752)
