+++ changelog.d/10752.misc
@@ -1,0 +1,1 @@
+Move tests relating to rooms having encryption out of the user_directory tests.
